Add files via upload
This update has the shuffle and betting functions. I used two random numbers to shuffle the deck array into a random order. Then I used a for loop to put this into a stack as the stack is best suited to the idea of dealing cards off the top of the stack. I also added a function to allow the user to place a Wager. I think I will run this after cards are dealt. The computer will automatically match the player's bet. The total wager from the computer and the player is displayed in the terminal once the bet is placed. Until this point I was still slightly confused about if my deck of cards had actually been made but now I can see it in the debugger menu which is great!