Add files via upload
This update has two methods that calculate the value of the players hand the computers hand. They are switch case based and also call a method to decide if the computer wants to play their ace as a 1 or 11 and lets the computer decide the same thing by entering 1 or 11.