Add files via upload
I was pre-mature to call the last version “Final” This version has a bug fix in the compare() method that should prevent false wins by the computer and has allowances for a tie. In a tie the player is the winner. Additionally the instruction.txt file has been renamed to Welcome.txt, partially in an attempt to diagnose why instructions stopped working in Visual Studio. However it does appear to work in J Grasp so I will submit it nonetheless.