Add files via upload
This update has a search method for searching the list by name. It currently has an issue and throws an exception.
This update also has the menu method from previous versions and a method that should be able to sub-group each data point by groups of 5. Using this we will be able to print only the desired data entries that combine to a complete phone book entry ( First Name, Last Name, Number, Street, City).